backend/requirements.txt
Django>=4.2
djangorestframework
pyotp
cryptography
argon2-cffi
python-dotenv
dj-database-url
psycopg2-binary
django-cors-headers>=4.0